Rename executable to trashy (#104)
Since trashy no longer supports `trash <PATH> ...` and has given up on
being a drop-in trash-cli replacement, the executable name `trash` is
only a hinderance to greater adoption because it conflicts with the
trash-cli executable of the same name. Giving the executable a unique
name allows both programs to be installed without issue and also allows
programs that wish to shell out to external commands for trashing files
to support trashy by simply checking if an executable named `trashy` is
present. Before this change, these programs had to check the output of
the user's `trash` command to determine if it was from trash-cli or
trashy.